requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
pyyaml==6.0.2
paramiko==3.5.0
